klippy: make MotionToolhead standalone, delete toolhead.py
MotionToolhead subclassed the legacy ToolHead to inherit the
extras-facing API during bringup, which meant the entire legacy
constructor ran at startup: LookAheadQueue, flush/priming timers
(immediately silenced), drip-move state, and junction-deviation
bookkeeping that no code path could reach.

MotionToolhead now owns its full surface directly:
- Move shrinks to the validation contract its remaining consumers use
  (kin/extruder check_move: axes_d/axes_r/move_d/end_pos, limit_speed,
  move_error); the junction/lookahead fields had no readers left
- register_lookahead_callback fires immediately with
  get_last_move_time(), which is exactly what the inherited version did
  against the permanently-empty lookahead queue
- limit_next_junction_speed keeps its empty-queue no-op semantics
- note_step_generation_scan_time keeps the flush side effect; the
  kin_flush_delay it used to maintain had no remaining readers
- junction_deviation / max_accel_to_decel / _calc_junction_deviation
  dropped: nothing reads them once Move stops tracking junctions
- SET/RESET_VELOCITY_LIMIT lose the kin.max_* branches BridgeKinematics
  never satisfies and keep the bridge.update_limits push

tools/test_motion_toolhead_static.py now guards the new invariants:
standalone class, legacy module absent, extras-facing API present,
no internal caller of the note_mcu_movequeue_activity no-op.
